examples/python/requirements.txt
grpcio==1.23.0
protobuf==3.7.0
six==1.12.0
